backend/requirements.txt
flask==3.1.0
flask-cors==5.0.0
google-genai>1.18.0
pillow==11.0.0
requests==2.32.3
python-dotenv==1.0.1
werkzeug==3.1.3
numpy>=1.26.0,<2.0.0
opencv-python==4.10.0.84
ultralytics==8.3.55
torch>=2.0.0
torchvision>=0.15.0
urllib3<2.0
